Introduce native YAML validator
Currently we use `kwalify` to validate YAML files, which is suboptimal
for several reasons: low performance, troublesome extensibiliy, and lack
of integration.

The only proper fix to that issue is to write a native validator from
scratch.  Given that we rely on YAML files heavily, this would be a
worthy investment.