caos: Namespaces -- Add support for namespace syntax to CaosQL
Caos semantic networks are expected to contain a large number of
concepts, thus the namespace will quickly become contended and messy.

To avoid that, introduce the notion of semantic namespaces where parts
of the whole semantic graph reside in their own namespaces.

This commit adds support for namespace syntax to CaosQL in the form:

   USING
       <namespace-name> [AS alias]
   SELECT
       (<namespace-name>|<alias>):concept

No actual logic is added, just the necessary modifications to the
grammar