caos: Implement namespace support
Having all the concepts and atoms in one big flat namespace is obviously
a bad idea.  So from now on each Caos schema defines a namespace.  That
namespace is actually a full name of the module the schema is defined
in, so this follows Python semantics, and furthermore, inside Caos
namespaces are called `modules`.

For example, a concept 'object' defined in semantix.caos.test.testdata
will have a fully qualified name 'semantix.caos.test.testdata:object'.
I chose to use a colon as a separator to avoid ambuiguity in CaosQL
where dot is used as a link element.

On Postgres side namespaces are defined as database schemas