caos.backends.pgsql: Turn 'entity' table into a root of inheritance
Previously the 'entity' table has been separated from the concept table
inheritance tree and simply stored entity ids and their corresponding
concept ids.  This is suboptimal for two reasons: a copy of every single
entity id in the database is duplicated in a single table, and it
introduces a potential need for an extra join to determine the concept
of the entity.

Fix this by placing the 'entity' table at the root of the inheritance
tree, so that every concept data table inherits from it.  This
eliminates the extra join and creates natural partitioning over
concepts.  The side effect of this is that direct foreign keys to entity
are no longer valid since the id index is no longer continuous.  For the
same reason entity id must be generated explicitly by calling the
entity_id sequence.