caos: Introduce a more Pythonic way to access Caos realm
Instead of providing clumsy factories with hidden magic configuration
expose Caos realm structure as a tree of objects that is constructed on
demand using backend metainformation.

So, instead of ugly

    factory = realm.getfactory()
    country = factory('semantix.caos.tests.testdata.geography_schema.Country')

comes shiny new

    with realm.root.semantix.caos.tests.testdata.geography_schema as sch:
        country = sch.Country

Using context manager is purely optional, but is recommended to improve
code readability.

Factories are still available, but are deprecated and used in places
where object class names are generated dynamically.

This is another step in the effort to further pythonify Caos interfaces.

During the course of implementation of aforementioned changes it was
uncovered that extracting meta from PgSQL backend was broken for some
time, but masked by the fact that YAML-originated metadata was used in
PgSQL tests.  This commit fixes that bitrot and restores proper function
to both PgSQL backend and its tests.