caos: Make objects behind one-to-one links accessible directly
This changes the behaviour of creating LinkSet regardless of link
mapping into that of always mapping linked objects directly if the link
has a one-to-one mapping.  This, by definition, affects all atoms.

This is an important change since it highlights the semantic difference
between one-to-one links and other links, besides making code more
readable.  Code breakage on mapping changes is also intentional and
beneficial, since mapping significantly changes semantics.