caos: Refactor and improve Caos Python binding
This brings a vast refactoring of Caos internals that brings a
wholly renovated Caos Python binding.

This rework completes the metadata/data duality whereby it is possible
to work both on a class and object level in the same way.

Concepts were given two fundamental methods to perform queries on the
graph: filter() and select().

filter() is mostly tailored for class-level use, where it is necessary
to fetch concepts concepts of a certain class given a path criteria, i.e
where the specified paths beginning at the concept yield non-empty sets.

    with semantix.caos.tests.testdata.geography_schema as geo:
        # Select all large European countries
        eu = geo.Country.filter(geo.Country.geography.continent.name == "Europe",
	                        geo.Country.geography.area.total > 300000)

On the other hand, select() given a specific set of links returns nodes
accessible via those links from the current node

    with semantix.caos.tests.testdata.geography_schema as geo:
        france = geo.Country.filter(geo.Country.name == "France")
	french_eth = france.select(france.people.ethnicities)

Both filter() and select() methods return a Selector object which is
evaluated lazily and can be chained.