caos: Rework Python binding transformer for more flexibility
Python binding's SelectorTreeTransformer is now stateful and Selector is
a thin wrapper around it.  The implementation of the transformer has
been reworked to allow more flexible and _iterative_ path construction and
target set selection.

Relying on the improved infrastructure is the ability to specify the
direction of the link in selector expression, e.g:

    SomeObject.select(caos.link.inbound(builtin.link))

This will select all objects that link _to_ SomeObject.