caos.pgsql.transformer: Implement more efficient atom reference resolution
Originally only id and concept_id fields were pulled up the select path
and for other atom references concept table needed to be joined at the
place of reference.  This was grossly inefficent as the concept table
was scanned at least twice as necessary.

Fix this by recording references to atoms and pulling them along id and
concept_id.  This removes the need to join concept table at
significantly simplifies atom reference resolution code, since it is now
safe to assume that the relation, containing the required atom is always
available in a given (sub-)query.