caos: Rework query engine
This brings an almost complete rework of Caos query engine.  The main
focus was reorganizing the expression processing to support alternative
and weak paths and, in general, making sure that the resulting paths
match the source query better.

To achieve that, the path processing is now done in two steps: building
the tree leaves-to-root and postprocessing the tree root-to-leaves.
Unlike previous implementation, no attempt of merging paths is made during
the first pass even if it was specifically requested by a common anchor.

Instead the second pass takes into account the semantics of the expression
the paths are in and combines them accordingly.

This work also brings the various tree transformers closer together and
factors out as much common code as possible.  The most noticeable change
is that CaosQL now uses link names to specify paths, instead of using
concept names.  This brings CaosQL in line with Python interface and, in
general, is a more logical syntax.  Also, to retain former
functionality, two new operators were introduced: IS and IS NOT, which
can be used to filter entity sets by the class.

Also, during this development, it became apparent that the current
implementation of AST infrastructure is too loose, so I took the
opportunity and reworked the generic AST field definition and
processing.