caos: Drop path cache infrastructure remnants
Path cache infrastructure was introduced, probably, with the first
iteration of Caos.  The purpose of it is to cache entity sets produced
by predetermined static paths.  The approach was very explicit and
inflexible.  Since then this code bit-rotted and is, most likely,
totally broken now.  Some sort of path caching will, inevitably, be
reintroduced but as an integrated part of Caos query engine.