caos.backends.pgsql: Modularize metadata synchronization
This lays the groundwork for metadata updates support.  Instead of
hardcoding the DDL operations, encode them as a set of
SynchronizationOperation objects, which is built on a metadata _delta_
which is calculated independently of the backend.  Right now the delta
is always the full metadata set, so this commit does not change the
functional behaviour.  Additionally, the backend now tries to import and
initialize all Postgres contrib modules it needs.