caos: Initial shot at incremental schema synchronization support
This is an initial proof-of-concept support for incremental schema
synchronization support for Postgres backend.  In other words, the
backend no longer expects an empty database when doing synchroniza-
tion.  It uses the results of generic schema comparison to perform
the necessary adjustments to the database.

This commit adds handling of most simple cases of incremental schema
changes, like atom mod and type changes, addition and removal of
links in concepts etc.

Currently, the synchronization is done in unattended mode, but the
plan is to make it interactive with heuristic analyzer merely giving
suggestions, charging the developer with the ultimate decision.
This is because schema changes are inherently dangerous and potentially
destructive operations, and there is no general way to automatically
devise a correct update plan in every case.