caos: incremental sync: Add rudimentary support for rename detection
Object renaming detection is, probably, one of the more difficult tasks
of heuristic schema analyzer.  To get better accuracy in those cases,
introduce a notion of context into the comparison process.  Now, when,
for example, two links are compared, it is possible to check if this
comparison is done in the context of the higher comparison of concepts
and discard the difference between the source ends of each link if those
ends correspond to concepts being compared.

This and minor tweaks to Postgres sync backend to support object
renaming make it possible to perform trivial renaming.