caos: Make sure that metadata checksum is stable
By definition we expect that the checksum is reliable and stable.  We
continue to use Python's hash() function, but make sure that it is used
only on 'stable' objects, i.e those with deterministic hash() value.
Thus, we rely on an interpreter implementation's promise to deliver
stable hash results for things like str, int, tuple and frozenset.  This
may turn out to be an over-extended assumption, but the current solution
should be good enough for now.