caos: Use persistent stable hash instead of Python's hash()
After some discussion it has been decided that Python native hash is far
too unpredictable to be used for an external persistent hash even in the
short term.  It simply was not designed for that use and never made any
such stability guarantees.  One can only rely on the stability of the
hash within one interpreter session.

Ergo, a persistent hashing algorithm was implemented and used instead.
To avoid reinventing the wheel badly I borrowed the implementation of
hashes for tuple and frozenset from CPython.