caos: Refactor meta sync plan generator to rely more on delta
The ultimate goal is to make the plan generator to rely only on
information from delta and not on prototype objects themselves.

With current delta implementation it is not possible to do that
separation completely since it lacks sufficient context information.
With this commit, though, it becomes easy to tell when and what
context is required.