caos: Use pymod: instead of yaml+file:// in backend URIs
yaml+file: never really made sense since the rest of the URI is
interpreted as a Python module path and is imported directly.
Furthermore, the whole point of hiding the representation language
behind the import is to use those modules without requiring explicit
knowlege of their format.

Thus, backend URIs containing module names must now use format-agnostic
pymod: schema.  The module is then loaded with generic Caos import
context and is examined to determine the language of the module to
select proper backend.

This commit also (temporarily?) disables support for multi-document YAML
modules, since the former syntax cannot be used with generic module
names and proper support will require some refactoring of the import
Finder.  Supplant this functionality with the ability to specify the
name of the module to use to qualify non-qualified names in the YAML
document itself.