utils.shell: Adjust command declaration, introduce MainCommand and command requirements
Command options must now be specified in the class keyword arguments
instead of the class dict.

'expose' class argument now acts by attaching the command as a
sub-command of the 'main command', which is defined by specifying
'__main__' as a command name.  There can be only one main command, but
it is possible to declare a subclass of the default MainCommand and that
subclass will supplant the main command.

Additionally, it is now possible to specify an arbitrary set of
command requirements: callables that raise an exception if the
requirement they denote is not satisfied.  Requirements are evaluated
prior to command execution.