Introduce Application class
For Semantix the meaning of an Application is a primary context of
execution.  Application specifies general configuration, Caos realm
setup, services to launch, etc.

This commit implements Application as a regular context manager that
emphasizes that an Application is a context.  Only one Application can
be active at one point in time.

Additionally, define a command requirement class -- ValidApplication --
that commands can use to signify that they require a valid application
context to run.