utils.lang.yaml: Adjust representer protocol
Make it clear that represent() method can be passed pure instances
of wrapped classes, i.e not the instance of the wrapper.  Thus,
represent() is turned into a classmethod with wrapped object being
passed as an argument.

Additionally, add an option to opt-out of Dumper's automatic aliasing
by passing 'ignore_aliases=True' to the wrapper class constructor.