utils.lang.yaml.validator: Implement 'class' schema type
'class' schema type is a mechanism useful when the type of the object
and rules of validation cannot be specified in the schema directly, e.g.
when the number of possible object types is large or unknown.

'class' type expects the following construct:

    fully.qualified.class.name:
        classfield1: value
	classfield2: value
	....

Known common fields can be specified in the schema directly:

    somekey:
        type: class
	fields:
	    classfield1:
	        type: str

But most of the time the class will implement `get_yaml_validator_config'
classmethod to expose the validation information for its fields.