caos: Initial implementation of metadata version control
This rework advances metadata synchronization between backends beyond
simple one-shot all-or-nothing process and brings us closer to the goal
of having powerful version control mechanism for Caos metadata.

Deltas are now formally managed by the 'metadata delta repository'.
Each backend has its own instance of delta repository, and the
implementation is also backend-specific (although, that is not strictly
necessary).

The process of metadata synchronization can now be described as a
synchronization of delta repositories which produces a set of deltas
which need to be applied to the backend to bring it in sync.

The organization and architecture of the delta repository has been
inspired by modern DVCSs (notably git).  This initial implementation
lacks any features beyond simple 'fast forward', though.

A number of deficiencies of backend metadata representation has been
fixed in the process and it is now mandatory that metadata hashes are
the same for all backends in the realm after synchronization has been
performed.