caos: Add creation/modification timestamps to the default concept base
caos.builtins.Object concept, which is the default base for all
concepts. got 'ctime' and 'mtime' atoms for creation and modification time
respectively.  The timestamps are updated automatically.

Where the presence of those timestamps is not desired, the concept can
be derived from caos.builtins.BaseObject which is the ultimate root of
the concept hierarchy.