caos.delta: Handle domain type changes properly
The change of the underlying data type for domains is a complex problem.
There is no direct way in PostgreSQL to change the base type of a domain.
Instead, a new domain must be created, all users of the old domain altered
to use the new one, and then the old domain dropped.  Obviously this
recurses down to every child domain.