caos: Add rudimentary Session implementation
Session is the ultimate abstraction of Caos transaction, or Unit of
Work.  In short, Session is a staging area, where changes to Caos
entities accumulate and can be later committed into the backend database
or discarded.

This is intentionally a very basic and rudimentary implementation that
does very limited change tracking.  Essentially, the functionality is
the same as before with Realm having an active default session, albeit
defined explicitly.

Entity cache has been moved out of realm and into the session object.
Non-root class cache is now also session-local as the session is
recorded in class metadata.

A new session can be creating by calling new_session() method of the
realm object.  When used as a context, session acts like a transaction
block, otherwise each database operation is automatically wrapped in a
transaction ('autocommit' mode).

It is possible to nest session transaction context:
   with session:
      do stuff
      with session:
          do nested stuff

Another notable change is that realm class tree can now be accessed
via realm 'schema' attribute instead of 'root'.  This is because
realm.schema is now an alias to realm.session.schema, since classes are
now bound to sessions.