utils.lang.import: Add support for lazy module loading
DSL module loading is fairly expensive and is not currently cached.
This noticeably affects startup performance.  To remedy that, make it
possible for modules to be loaded lazily when they are first used (an
attempt to access an attribute of a module is made).

This gives us a win of ~100ms of startup time.  Unfortunately, lazy
loading breaks scenarios where no objects are actually referenced from
the module and its loading is used to generate some side-effect
(e.g. populate config), so keep it off for now.

Additionaly, make some tweaks to conform to PEP 302 better.