caos: Implement entity cache and entity expiration
Entity objects cached in session are now automatically reused when they
appear in the result set of a selector.  Entities can now be marked as
expired.  If an entity is expired, further attempt to access the value
of its link will cause that entity to be reloaded from the database.

To expire an entity one should call session.expire(entity[, links]),
where links is an optional iterable containing names of the links that
should be expired -- it is possible to expire an entity partially.  By
default all links are expired.

There is also a way to expire all entities in session in bulk by calling
session.expire_all().