caos.backends.yaml: Adjust multi-target link syntax
Previously when a multi-target link was defined and there was a need to
override its properties one had to list each target as a separate key in
a mapping and repeat the properties for each item:

    multilink:
         target1:
             mapping: "1*"
         target2:
             mapping: "1*"

This is not only inconvenient, but illogical, since the properties for
each target must be the same.  The following syntax must be followed
instead:

     multilink:
         [target1, target2]:
             mapping: "1*"

This is a logical continuation of the

     multilink: [target1, target2]

syntax which was already present for links which did not need to
override any properties.