caos.caosql: Adjust the syntax to use '.' instead of ':' for module qualification
This is (hopefully) the solution for the long-standing problem of
conflicting use of dot as both a path operator and a namespace attribute
getter.  Previous attempt was to use a colon as a module:name separator,
but this had a number of severe shortcomings: poor readability, mismatch
with the rest of the system, and, most notably, it was only possible to
unambiguously use it with short module aliases, i.e module name could
not have dots in its name.

So, the solution is to use dot as a module.name separator, and require
enclosing such qualified names into square brackets to resolve ambuiguity
with path expression.  Square brackets were already used for link
expressions, so it is logical to extend their use for all path point
nodes.

Example:

     SELECT
          [my.ns.Concept].[my.ns2.link].[my.ns3.link2]