datastructues: Add an option of deferred initialization of defaults to Struct
In some cases it is not desirable to initialize empty Struct fields with
default values at the moment of construction.  Such "partial" Struct is
useful with inheritance and merging patterns.  To initialize empty
fields with defaults a setdefaults() method should be called.

Another effect of this change is that it is no longer possible to set a
field to None if that field does not have None as its default.