caos: Make sure that links prototype attributes are inherited properly
This touches predominantly the YAML backend.  Since objects in YAML can
be specified in partial form and set to inherit from other objects, care
should be taken so that defaults do not overwrite values that should be
inherited from parent.  Thus, a non-null default for prototype property
in YAML schema is almost definitely a bug, although no action is taken
in this regard yet.