caos: Escalate backend link mapping cardinality violation errors
When backend storage raises an exception turn it into a StorageError and
make the session handle this situation properly by going into a special
"crashed" state which prevents all further synchronization operations
until the session is rolled back and the backend storage is reset.

Note that the session does not take any action against the cause of
the error which is usually an invalid entity.  In other words, a broken
entity will remain in the session's dirty index and the error will be
raised again on repeated sync().