caos: Improve handling of constants
Constant expressions are now treated like constants, so, `2 + 2` is a
constant and `$var = 0` is a constant too.

Also, constant nodes are now annotated with the data type of the
constant, since it is known to the application in any case and storage
backend sometimes needs explicit type information to process queries
properly.