caos.backends.pgsql.parser: Implement parser for a subset of SQL
This adds a parser for an expressional part of PostgreSQL SQL, the part
that is required to parse index conditions, expression constraints etc.
Based largely on the PostgreSQL parser sources.

Word of warning: this grammar seems to be too much for Pyggy to handle
gracefully -- it takes approximately 5 minutes to generate parser
tables.