utils.ast: Implement basic AST tree pattern matcher
It is sometimes necessary to match an AST tree against some predefined
pattern and potentially distill some information from it, like the
values of constants in some expression.

This commit adds a simple AST patter matcher.  The pattern is built from
the AST nodes we are matching adapted to MatchASTMeta, so a tree

    ast.BinOp(left=ast.Constant(), right=ast.Constant(), op='+')

can be matched with

    astmatch.BinOp()

where astmatch.BinOp is an adapter for ast.BinOp.  Note that only the
fields explicitly specified in the pattern node are significant.  Fields
that are not specified will not be taken into account.

It is possible to specify alternative patterns with ast.match.Or so that

    ast.match.Or(astmatch.BinOp(), astmatch.Constant())

will match either a BinOp node or a Constant node.

Capturing groups are also supported.  To declare a capturing group,
enclose the node of interest into ast.match.group:

    pattern = ast.match.group('mygroup', astmatch.BinOp())

The groups can be nested: ast.match.group can appear within a node
being matched by another group:

    pattern = ast.match.group('mygroup',
                              astmatch.BinOp(
                                  left=ast.match.group(
                                          'constant',
                                          aastmatch.Constant())
                             ))

And last, the actual matching is done by calling ast.match.match:

   result = ast.match.match(pattern, tree)

If there was a match, the result will contain a MatchGroup object with
'node' attribute pointing to the matching node.  If there were any
capturing groups defined, their contents will be a list of MatchGroup
objects available through the MatchGroup attributes corresponding
to the group names:

   >>> print(result)
   <semantix.utils.ast.match.MatchGroup object at 0x7f507ac9f350>
   >>> print(result.mygroup)
   [<semantix.utils.ast.match.MatchGroup object at 0x3620e50>]
   >>> print(result.mygroup[0].node)
   <myast.BinOp object at 0x34314520>
   >>> print(result.mygroup[0].constant[0].node)
   <myast.Constant object at 0x3414810>