caos: Implement text search configuration infrastructure
This provides the necessary foundation for text search support.  Links
gained a new 'search' attribute, which, when not empty, instructs Caos
to perform necessary actions to facilitate text search in the node
pointed to by the link in context of the source.  Currently, only atom
targets are supported.

The 'search' attribute of the link may be set to either True, False or a
mapping containing more detailed search configuration.  Currently, only
one search configuration attribute is recognized: 'weight', which
specifies the relative importance of the contents of the link target in
the overall ranking.

Note that this commit contains no attempts to use the new
infrastructure, there are no provisions for text search clauses in Caos
queries yet.