caos: Simplify root link name extraction
Specialized (derived) links used to assume that their direct parent is a
generic link and tried to use first parent's name unconditionally.  This
assumption is not safe, since link inheritance hierarchies can and will
be deeper, so the correct way is to obtain a normal_name() of the link
and fetch the generic link prototype directly from the meta index.