caos: Get rid of dubious 'implicit_derivative' link notion
The original idea was that each concept link implicitly inherited from a
global link definition with the same name.  Even though such implicit
inheritance does take place, having a special 'implicit_derivative'
property denoting that fact is superfluous.

The real conceptual separation is that there are two kinds of links.
First, there are "generic" links -- the ones that are defined on global
level, have definitions of properties etc.  And second, there are
"specialized" links -- the ones that are defined within the context of
concepts and have specific source and target.