caos.backends.pgsql: Throw away pyggy and use Parsing for SQL parser
Aside from the fact it was taking more than 5 minutes for pyggy to
generate parsing tables, it just doesn't work on complex grammars.  No
diagnostics, no proper precedence control, random bugs all made it
unbearable.  So I threw it away and reimplemented the parser using
Parsing.

Now, Parsing isn't exactly an angel too, but at least it has good
built-in diagnostics, _and_ checks the grammar _at compile time_
(for sanity's sake, pyggy, how hard is that?).

In terms of performance, after my tweaks to Parsing, it takes 1:40 to
generate the parsing tables.  Not exactly stellar, and I'm sure there
are more opportunities to bring the time down.  In the end, though, I
think it would make sense to include the compiled tables in the
repository.

Oh, and pylly is still used for lexing.  We'll probably need to rewrite
it to ply.lex, although, I'm quite skeptical that it will be a
pleasant experience.