caos.backends.pgsql: Use table constraints instead of domains for automatic atoms
The original approach was to use database domains for automatic atoms
too.  While techically correct, this approach has one major showstopper:
PostgreSQL requires that corresponding columns in child tables must be of
the exact same type as columns in parent table.  This makes use of
domains for automatic atoms impossible.

There is a workable alternative: automatic atom mods can be represented
as table constraints.  This way, the base type stays the same and proper
inheritance is possible, since table constraints have cumulative properties,
i.e child table will have the sum of table constraints of all its parents,
which is exactly what we need.

Another important aspect of this change is that we no longer translate
max-length mods into database type length modifiers.  Again, the reason
is the same: base column types must match.  What this means is that
we no longer use varchar(n), all string columns are defined as 'text'.
Same applies to numeric.

It is worth mentioning that, while originally perceived to be
significant, this change brings very little impact on performance, if
any.  For reasons why unbounded "text" is just as good as "varchar(n)"
in PostgreSQL see http://tinyurl.com/pgtext

For numerics too, remove all the magic and use plain 'numeric' every
time, but for builtins.decimal atoms _only_.  builtins.int now maps to
database BIGINT and is artificially limited to the 64-bit range:
-0x7FFFFFFFFFFFFFFF to 0x7FFFFFFFFFFFFFFF.  It will probably make sense
to map decimals with scale zero to regular unbound ints, but it doesn't
seem to be worth the effort at this moment.