caos.tree.transfomer: Adjust inline filter logic
Commit '8ca0cf9 caos: Do not try to fold complex expressions
into inline filters' made it so only first-order atomic reference
expressions were inlined.  This obviously broke all higher-order
expressions, like '(ref / const1) > const2'.  The original change was
done so that expressions like '$const is None OR some.ref = $const'
worked as expected, but the proper fix would be to mark that the need to
evaluate 'some.ref' path depends on the value of the constant, but only
in '<constval> OR <path>' kind of expressions.