caos.expr.transformer: Simplify Caos query tree assembly
Previous coding had to process the same sub-expressions twice to build
the query: first to extract paths, and then to assemble the final tree
after the paths have been processed and merged.  With good use of
processing contexts it is possible to do expression interpretation
only once and then do simple fixups to the processed expressions after
path merging.