caos: Add initial support for aggregates in queries
This adds initial support for aggregates to Caos queries in both
expressional and CaosQL forms.  The semantics of aggregation is similar
to that of SQL.

CaosQL gained the GROUP BY clause and the ability to use agg: functions:

        SELECT
            E.name                  AS ethnicity,
            agg:count(Country)      AS no_countries
        WHERE
            Country.people.ethnicities{E} in %
            AND agg:count(Country) > 2
        GROUP BY
            E.name
        ORDER BY
            agg:count(Country)

The above query will fetch the names of ethnicites and the number of
countries that have them if that number exceeds two.

Note that unlike SQL there is no separate HAVING clause to place
restrictions to the aggregated values.  Instead, those restrictions
can be written in the WHERE clause along with regular restrictions.
One obvious requirement is that it must be possible to factor out
the aggregate restrictions, i.e:

    WHERE (aggregate restrictions) AND (regular restrictions)

Thus, the following is invalid:

    WHERE agg:count(Country) > 2 OR Country.name = 'Germany'

On the Python Caos expressions side Selectors learned the 'group()'
operation, which has the same effect as the GROUP BY clause.
As always, the semantics is the same as in CaosQL:

    Ethnicity.query(
                 ethnicity = Ethnicity.name,
                 no_countries = caos.expr.agg.count(Country)
              ).filter(
                 caos.expr.agg.count(Country) > 2
              ).group(
                 Ethnicity.name
              ).sort(
                 caos.expr.agg.count(Country)
              )