utils.ast: add LanguageAST base class.
This class adds the ability to track information of where an ast node
came from by adding one more field to the ast node class:
'source_position'.

Will be used extensively in various implementations of custom languages.
Now it is used as a base class for Python and Generic language asts.