utils.lang.generic: Add 'generic' language.
Generic Language is a simplified and generalized version of Python
Language.  For instance, 'a, b = 1, 2' is one AST node in Python, and
two separate Assignment nodes in Generic.  Hence, it is much easier to
write code analyzers and reflectors using Generic language
infrastructure.  It can easily be converted to python back, which is
actually done in unittests to ensure translation correctness.

Right now, a very limited set of python language features are
implemented: functions (arguments and return statement), basic
operations (binary, boolean, unary), and literals.