caos.backends.pgsql: Fix handling of long object names
The length of practically all object names in PostgreSQL is limited to 63
characters.  At the same time, it is quite possible to get the name that
exceeds that length in Caos.

The mechanism to trim names by replacing the necessary chunk of the name
with base64-encoded hash of the name was implemented a while ago,
however, it was not used consistently.

This change brings consistent handling and translation of Caos names to
PostgreSQL object names.  All attempts to infer Caos name from
PostgreSQL object name are invalid and have been removed.  Name mappings
have been introduced where necessary.

Additionally, bring some love to link properties: create meta table for
link properties and be more careful with link property names.