caos: Improve handling of atom and link defaults
The mechanism of default values for atoms and links has got some
treatment.  First, it actually works properly on merged links now, and
second, a mechanism to specify a list of defaults instead of just one
has been introduced.

The idea is to have more than just static literal defaults, but to also
allow arbitrary queries to obtain default values for atoms and link
targets, and, additionally, have a mechanism to fallback to a static
value if the query fails.  Having a queue of options to fetch defaults
from also streamlines inheritance -- defaults can simply be appended.

This commit brings the necessary infrastructure for different kids of
defaults, but does not actually implement any of them besides the
LiteralDefaultSpec, which is the old literal default.