caos: Make link property a first class Caos object
Link properties are in many ways very similar to atomic concept links,
and there is no reason why the former should be treated in some special
way.

Conceptually, there are three main functions for every top-level
Caos object: Source, Pointer and Target.

There are three possible combinations:

Concept -- Link -- Concept, Concept -- Link -- Atom, and
Link -- Link Property -- Atom

So Link objects function both as a Pointer and as a Source.  In light of
this there is a potential to reuse large portions of Link-as-pointer
implementation for link properties.

This commit implements the necessary backend metadata infrastructure.
No changes at the interface level, but the plan is to exhibit Source
behaviour for Links, i.e., expose link properties as class attributes,
etc.