caos: Cast atom mod values to the proper type before adding them
The type of certain Atom mods' values must correspond to the data type
of the atom they are attached to.  For example, MaxValue/MinValue family
has that requirement, since they do direct comparisons against atom
value.

This is also a good opportunity to get rid of the remnants of old
regexp-based constraint interpretation in PgSQL backend and replace it
with AST expressions.