caos: Bring link properties functionality to substantial completness
Link properties are now substantially feature-complete.  Every aspect of
link properties handling was brought to the level of concept-link
functionality: proper schema handling with inheritance and
specialization support, support in queries, session tracking etc.

A lot of common code has been factored out and is shared by both
Concept-Link-Target and Link-Property-Atom under the
Source-Pointer-Target generalization.

It is now possible to obtain the link object for a particular link with
the caos.concept.getlink(entity, link_name[, target]) function.  The
link properties are exposed as link object attributes just like links
are for concepts.  References to link properties are accepted by the
Python Caos query expressions just like link references are.

In CaosQL it is now possible to reference link properties using the "@"
indirection:

    SELECT
         Country.president@from
    WHERE
         Country.president@to < $date

Here, "from" and "to" are properties of the "president" link.