caos.backends.pgsql: Check for base atom type in the "adapts" class attribute
Custom atom base classes should now define the name of the built-in atom
they "adapt" in the "adapts" class attribute.  That is, if the custom
base class extends from str, "adapts" should contain
"semantix.caos.builtins.str".

This is considered to be a temporary workaround until a more elaborate
schema to support custom atom base classes is developed.