caos: Remove the need to pass a full-blown realm to the CaosQL transformer
In a number of places the requirement to pass a full-blown realm was
superficial, since the only use of it was to access the realm prototype
schema.  Additionally, a number of methods in the prototype schema
itself requested the realm object in the same manner.  The
prototype schema object is now passed directly and, thus, the
requirement of having an initalized realm is no longer.