caos: Implement basic provisions for expression type inference
Caos now knows a little bit more about the data types in expressions and
describe_output() method of the query statement now returns an actual
Caos type where possible.  Additionally, it is possible to obtain a
list of query arguments and their types by calling describe_arguments()
method of the prepared query statement.

It is worth noting that the type inference framework is very incomplete
and is only capable of handling only a small subset of expressions, but
it is a good start.