caos: Fetch and construct entity objects in one shot when possible
Previously a reference to a Concept in the query result expression would
have been transformed into an id reference and then the caller would
have had to manually load atoms in a follow-up query, which resulted in
at least two, and sometimes three, queries to load an entity.

With this change a reference to a Concept is transformed into a row
expression that selects all atoms of a matched entity.  This row is then
processed by a custom RowConstructor to create an entity object.  In
other words, only one query is needed to load an entity and its atoms.
However, this is only possible when the selector reference resolves
to a single concept, otherwise it is not possible to determine the exact
type of the row type and a fallback to previous behaviour of selecting
an id and concept name is made.

Note that this change requires small modification to py-postgresql which
makes it pass the relation oid to a row constructor factory.