caos: Implement batch session sync mode
When a number of changes in the session is large it is much more
efficient to commit them to the persistent storage in batch mode.

It is now possible to instruct the session to perform sync() in batch
mode by creating a batch context:

    with session.batch():
        <code>
	session.sync()

When in batch context, sync will write data to the persistent storage in
bulk instead of one-by-one.

For now batch mode can only be triggered explicitly as shown above.
Automatic triggering by session changeset size heuristics is a
possibility, but I will leave it for later.