caos: Implement session pooling
Sessions are isolated unit-of-work containers and require careful
management in multiuser environment to enforce strict separation and
resource allocation control.

This changeset introduces a notion of session pool.  If a service
request handler needs a session, it requests it from the session pool
and releases it when it is no longer needed so that it can be recycled
and used again by another request.  Connections to persistent storage
are pooled in the same manner.  For now, sessions hold a connection for
their entire lifetime, though.

Additionally, the dubious default realm session has been killed, since
sharing a common implicit session is a very bad idea.