caos: Implement "sequence" builtin
Sequence is a generic counter atom that is guaranteed to return a unique
sequential value on each creation.  The counter value can optionally be
wrapped into an arbitrary text pattern which can be specified in the new
"attributes" dictionary of atom definition.

It is possible to set a magic default -- "next" -- for sequence atoms
which will instruct the atom to generate the next value of the sequence
if an empty value has been passed into its constructor.  Same effect is
achieved by passing "next" into constructor.