caos: Use the whole inheritance tree to determine the existense of a path
Previously a reference to a pointer of a certain source would fail if
that pointer was not immediately defined in the source or one of its
parents.  Now Caos would also consider all children of the source to
determine the existense of the path.  Consider the following example:

    User:
        links:
	    login: str
	    authentications: UserAuthentication

    UserAuthentication:
        abstract: True

    PasswordUserAuthentication:
        links:
	    password: str

    FingerprintUserAuthentication:
        links:
	    fingerprint: str

Previously, the following expression would have been invalid:

    User.authentications.password == 'something'

But now the "password" pointer will be found in the
PasswordUserAuthentication child and that path will be selected,
effectively transforming the expression in to the following CaosQL
expression:

    User.authentications{A} in %
    and A is PasswordUserAuthentication
    and A.password = 'something'

In other words, Caos now considers a much larger set of paths to
determine paths that can possibly match the query.